+++ packages/lapikit/src/internal/core/formatter/index.ts
@@ -1,0 +1,33 @@
+import { preset } from '$lib/internal/config/presets.js';
+import { deepMerge } from '$lib/internal/deepMerge.js';
+import type { DevConfiguration } from '$lib/internal/types/index.js';
+import { componentFormatter } from './component.js';
+import { themesFormatter } from './theme.js';
+import { typographyFormatter } from './typography.js';
+
+export async function css(config: DevConfiguration) {
+	// states
+	const defaultTheme = config?.theme?.defaultTheme || preset.theme.defaultTheme;
+	const defaultTypography =
+		config?.typography?.defaultTypography || preset.typography.defaultTypography;
+
+	// formatter
+	const themes = themesFormatter({
+		themes: deepMerge(preset.theme.themes, config?.theme?.themes || {}),
+		defaultTheme
+	});
+
+	const typography = typographyFormatter({
+		typography: deepMerge(preset.typography.fonts, config?.typography?.fonts || {}),
+		defaultTypography
+	});
+
+	// components
+	componentFormatter({
+		breakpoints: deepMerge(preset.breakpoints.thresholds, config?.breakpoints?.thresholds || {})
+	});
+
+	return {
+		styles: [themes, typography]
+	};
+}