+++ packages/lapikit/CHANGELOG.md
@@ -21,6 +21,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2025-08-31
+
+### Added
+
+- Store **breakpoint** for use _mediaQueries_ on your application
+- Store **devices** for use breakpoint on viewport
+
+### Changed
+
+- Update function for change _theme_ and _colorScheme_
+
 ## [0.2.2] - 2025-08-20
 
 ### Added